+++ example/middleware/auth.js
@@ -2,7 +2,7 @@ import { timingSafeEqual, randomUUID } from 'node:crypto';
 
 /**
  * Basic auth middleware
- * @param {Map<string, string>} users
+ * @param {Map<string, User>} users
  * @param {boolean} [allowAnonymous]
  */
 export function basicAuth(users, allowAnonymous) {
@@ -12,22 +12,23 @@ export function basicAuth(users, allowAnonymous) {
    * @param {import('express').Response<any, {user:User}>} res
    * @param {import('express').NextFunction} next
    */
-  return async function basicAuth(req, res, next) {
+  return function basicAuth(req, res, next) {
     const authHeader = req.get('Authorization');
     if (!authHeader) {
       if (allowAnonymous) return next();
       return sendUnauthorized(res);
     }
 
-    const auth = Buffer.from(req.get('Authorization').substring(6), 'base64').toString();
+    const auth = Buffer.from(authHeader.substring(6), 'base64').toString();
     const [username, password] = auth.split(':');
 
     try {
-      const user = await authenticate(users, username, password);
+      const user = authenticate(users, username, password);
       if (!user && !allowAnonymous) {
         return sendUnauthorized(res);
       }
 
+      // @ts-ignore
       res.locals.user = user;
       next();
     } catch (err) {
@@ -41,7 +42,7 @@ export function basicAuth(users, allowAnonymous) {
  * @param {Map<string, User>} users
  * @param {string} username
  * @param {string} password
- * @returns {Promise<User|undefined>} user
+ * @returns {User|undefined} user
  */
 function authenticate(users, username, password) {
   const user = users.get(username.toLowerCase());
@@ -58,6 +59,7 @@ function authenticate(users, username, password) {
     return;
   }
 
+  // @ts-ignore
   return { username, ...user };
 }
 
@@ -75,7 +77,7 @@ function sendUnauthorized(res) {
  * @typedef {Object} User
  * @property {string} username
  * @property {string} name
+ * @property {string} password
  * @property {string} [email]
  * @property {string[]} [role]
- * @property {string} [password]
  */